Switch to new config locations & write system
Use `~/.config/tv/*` instead of `~/.config/pdp/tv`, use default symlinks,
manage the ability to write to files.

Sets up vaguely-XDG compliant locations (no path walking, or fallbacks, but
works for the simple cases).

Uses pdpgeneric.VerboseObject, which I might need to revert since that's
supposed to be an optional dependency.  But I'm leaving it in for now to aid in
debugging.

Writing a new config needs a dummy command, such as `:help`; I might want to
use `:write foobar` instead of `--new-tvname foobar -w` to get around that.
Hum.

In the end, no new TV interaction features.

This sets the stage for adding RSA key generation and TV RSA key fetching and
caching, so that we can send/receive text encrypted, for entering passwords.