Default to a REPL if no commands given
This is using readline and supports a history file, and an input config file,
but is otherwise currently naïve.

But it works and is a bit easier to use for navigating menus.

Needs:
 * tab-completion
 * sub-menu for key input (morally same as '['...']'
 * better split of parameters, to support quoting for future text

Subclassed Exit with BadInput to better distinguish which errors are fatal and
which just need a re-prompt.