tv: tab completion, quoting, etc
Implement tab completion.  It's a mess, as tab completion with existing
libraries always seems to be.  Handle whitespace in app names.

Tab complete top-level commands, keys, and app names.

Fix the REPL commands to use `shlex.split` so we can quote things.

Allow comments in the REPL.  Exit gracefully on Ctrl-C.

Be more resilient to TV errors in a few places.