consistency in colon-mode plus key command
If there's a REPL, then the `key` command prompts for keys in a sub-REPL.  With
curses, in colon-mode if given a bare `key` then the resulting behavior varied
depending upon how we got into curses.  If direct, there was no REPL and we
gave an error message.  If via the REPL, then the key sub-REPL was used, and we
effectively switched _out_ of curses, instead of stacking.

Instead, we now evaluate the commands with the existence of any REPL
suppressed, to get consistency and avoid a broken stacking/switching model.